requirements.txt
asgiref==3.5.2
Django==4.1.3
django-environ==0.9.0
django-jazzmin==2.6.0
flake8==5.0.4
mccabe==0.7.0
psycopg2==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
sqlparse==0.4.3
